Fix CockroachDB decimal/float complaints and add retrying to loading

+++ src/com/oltpbenchmark/benchmarks/tpcc/procedures/Delivery.java
@@ -22,6 +22,7 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Timestamp;
 import java.util.Random;
+import java.math.BigDecimal;
 
 import org.apache.log4j.Logger;
 
@@ -216,7 +217,7 @@ public class Delivery extends TPCCProcedure {
             rs.close();
 
             int idx = 1; // HACK: So that we can debug this query
-            delivUpdateCustBalDelivCnt.setDouble(idx++, ol_total);
+            delivUpdateCustBalDelivCnt.setBigDecimal(idx++, BigDecimal.valueOf(ol_total));
             delivUpdateCustBalDelivCnt.setInt(idx++, w_id);
             delivUpdateCustBalDelivCnt.setInt(idx++, d_id);
             delivUpdateCustBalDelivCnt.setInt(idx++, c_id);
